+++ src/core/inventory/infrastructure/repositories/graphql/responses/inventory-item-find-by-id.response.ts
@@ -1,0 +1,5 @@
+import type { InventoryItem } from '@/core/inventory/domain/types/inventory-item.interface';
+
+export interface InventoryItemFindByIdResponse {
+  inventoryItemFindById: InventoryItem | null;
+}