+++ src/core/inventory/infrastructure/repositories/graphql/responses/inventory-items-find-by-criteria.response.ts
@@ -1,0 +1,5 @@
+import type { InventoryItem } from '@/core/inventory/domain/types/inventory-item.interface';
+
+export interface InventoryItemsFindByCriteriaResponse {
+  inventoryItemsFindByCriteria: { items: InventoryItem[] };
+}